TBTrampoline ARM64 complete
ARMv6/7s and x86 incomplete, unsupported currently.

It works by passing all register args and the stack as buffers to a landing funciton, which processes them to replace the arguments.

+++ TBTweakViewController/TBTweakViewController/Classes/TBTrampolineLanding.h
@@ -15,4 +15,4 @@ typedef uint8_t byte;
 #define kArgRegCount 8
 
 /// Function to prepare trampoline landing to receive arguments. FPRegisters may be NULL.
-IMP TBTrampolineLanding(id obj, SEL sel, byte *stackArgs, byte *GPRegisters, double *FPRegisters);
+IMP TBTrampolineLanding(id obj, SEL sel, byte *stackArgs, uintptr_t *GPRegisters, double *FPRegisters);
